trim some code from pkg/types
